feat(accounts): replace accounts screen with avatar MeMenu
Clicking the user avatar opens an account menu directly: header (avatar +
name + npub with copy), inline account list with one-tap switch and per
row rename/remove, plus add-account, settings, and logout. Removes the
need for a dedicated accounts screen and the Settings → Accounts entry.

Shell adapts to window width (threshold 912dp, matching App.kt):
  - Desktop: Popup anchored above the rail avatar.
  - Mobile / narrow: ModalBottomSheet.

Notifications header chip now opens the menu instead of navigating to the
deleted Screen.Accounts. The orphan sidebars/Sidebar.kt (no callers since
the desktop shell moved to ServerRail) is removed alongside.
